[Update] Remove main-column() mixin, use grid-row() instead
Foundation 5 has a max-width for rows set in _settings.scss. This is so
the page doesn't become unreadable (very long lines are hard to read).
This is exactly why we have the `main-column()` mixin. So we can just
use Foundation's rows instead, and set the maximum column width via the
$row-width variable in _settings.scss