[Fix] Add dry-run check in migrations
Migration 0006 and 0008 in the "skills" app was not working properly.
These migrations include data migration, which cannot be done on a dry
run, which is what happens the first time you migrate after syncdb.

The solution (suggested by `south` itself) is to wrap the data migration
in `if not db.dry_run:`.