Adds sanity tests
This adds to sanity tests to the grunt file and the javascript needed
for them. One of them, front-page-sanity-test, checks that the page
displays an element correctly (ie. its text is equal to a know string
specified beforehand) and the other checks that a known username and
password combination allows for login to the server. The username and
password are given to the program in the javascript file.

Updates the grunt file to add support for the new tests.

Adds a name label to an element in the front_page.html file to make it
easier to reference from the front-page-sanity-test.

Adds javascript files for both of the new tests.