Adds link test and splits sanity test into two tests.
Splits sanity test into two separate tests: one to test if login is
possible (login-test) and another to determine if the front page is
displaying properly (front-page-sanity-test). Also adds link-test, which
goes through all the links on the website it can find and checks to see
if any of them 404.

Adds file for link-test.

Splits sanity-test into two separate files.

Changes grunt file to allow testing with link-test.

Modifies the 404 template to make it easier to tell if a page 404d.