[MAJOR] Changes the tests from grunt to django
This commit changes the tests from working with grunt, which got very
messy, very quickly, and replaces it with the built in django test
framework. Also adds skill completion test, which tests that a skill
can be completed, and rewrites login-test to work with the new system.
Moves away from nose as well.

Minor changes to skills.json

Adds a name flag to an element of skill_view to make it easier to
find with selenium

Adds a name flag to an element of trainingbit_cover to make it easier
to find with selenium

Adds tests.py, the new file for tests to be written in.