Adds pdiff
This commit adds pdiff to the test file. It takes about a minute to run,
but works. The resulting image is saved as pdiff.png. There is also a
file called newscreenshot.png. This is automatically generated each time
the test runs, so it’s not important if it’s there or not.